player_frontend/requirements.txt
nicegui>=1.4.20
python-chess>=1.999
httpx>=0.27.0